# 🌍 #349 Performance improvements