--- a/src/GalaxyCheck/Gens/ListGen.cs
+++ b/src/GalaxyCheck/Gens/ListGen.cs
@@ -141,7 +141,7 @@ namespace GalaxyCheck.Gens
 
         public IListGen<T> DisableCountLimitUnsafe() => this with { EnableCountLimit = false };
 
-        protected override IGen<IReadOnlyList<T>> Get => TryGetCountRange(Count).Match(
+        public override IGen<IReadOnlyList<T>> Get => TryGetCountRange(Count).Match(
             fromLeft: range => GenList( ElementGen, range, Bias ?? Gen.Bias.WithSize, EnableCountLimit ?? true),
             fromRight: gen => gen);