Lots of code added that prevents reanalyzing reflection, or recreating gens.

Before:
```
|                   Method | iterations |       Mean |     Error |     StdDev |     Median |
|------------------------- |----------- |-----------:|----------:|-----------:|-----------:|
| SampleClassWith10Members |          1 |   1.109 ms | 0.0431 ms |  0.1258 ms |   1.070 ms |
| SampleClassWith10Members |         10 |   9.865 ms | 0.8243 ms |  2.4304 ms |  10.143 ms |
| SampleClassWith10Members |        100 | 101.228 ms | 4.0065 ms | 11.6237 ms | 102.502 ms |
```

After:
```
|                   Method | iterations |        Mean |     Error |    StdDev |      Median |
|------------------------- |----------- |------------:|----------:|----------:|------------:|
| SampleClassWith10Members |          1 |    959.2 us |  29.59 us |  86.77 us |    924.7 us |
| SampleClassWith10Members |         10 |  3,112.6 us | 108.18 us | 315.56 us |  3,066.1 us |
| SampleClassWith10Members |        100 | 22,662.2 us | 448.86 us | 926.97 us | 22,578.3 us |
```